Remove the fixture runner in favour of a normal junit test and add an additional variant test which verifies the main source set can query tables from other source sets
