--- sqldelight-gradle-plugin/src/test/fixtures/fixtures/failure.txt
@@ -1,4 +1,0 @@
-src/minApi21DemoDebug/sqldelight/com/sample/demo/debug/DemoDebug.sq line 7:5 - No table found with name full_table
-6    SELECT *
-7    FROM full_table
-          ^^^^^^^^^^
