+++ sqldelight-gradle-plugin/src/test/fixtures/fulfilled-table-variant/src/internal/sqldelight/com/sample/demo/SomeTable.sq
@@ -1,0 +1,4 @@
+CREATE TABLE some_table(
+  _id INTEGER NOT NULL PRIMARY KEY,
+  some_column TEXT NOT NULL
+);
