+++ sqldelight-gradle-plugin/src/test/fixtures/fulfilled-table-variant/src/release/sqldelight/com/sample/full/SomeTable.sq
@@ -1,0 +1,4 @@
+CREATE TABLE some_table (
+  _id INTEGER NOT NULL PRIMARY KEY,
+  some_column1 TEXT
+);
