+++ sqldelight-gradle-plugin/src/test/fixtures/variants/src/androidTest/sqldelight/com/sample/android/test/AndroidTest.sq
