+++ sqldelight-gradle-plugin/src/test/fixtures/variants/src/demo/sqldelight/com/sample/demo/Demo.sq
