+++ sqldelight-gradle-plugin/src/test/fixtures/variants/src/full/sqldelight/com/sample/full/Full.sq
