+++ sqldelight-gradle-plugin/src/test/fixtures/variants/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.example.sqldelight"/>
