+++ sqldelight-gradle-plugin/src/test/fixtures/variants/src/main/sqldelight/com/sample/MainTable.sq
