+++ sqldelight-gradle-plugin/src/test/fixtures/variants/src/minApi21DemoDebug/sqldelight/com/sample/demo/debug/DemoDebug.sq
