+++ sqldelight-gradle-plugin/src/test/fixtures/variants/src/minApi21FullDebug/sqldelight/com/sample/full/debug/FullDebug.sq
